Change MoveItem to Uint16 and fix the encoders/decoders for future use (only numbers 8 to 255 are currently in use).

+++ heromesh.h
@@ -320,7 +320,7 @@ void swap_world(void);
 
 // == game ==
 
-typedef Uint8 MoveItem;
+typedef Uint16 MoveItem;
 
 extern MoveItem*replay_list;
 extern size_t replay_size;
